==> trial/scripts/00_Source.R <==
library(tidyverse)
library(mgsub)
library(ochRe)
library(ggforce)
library(ggpubr)
library(ggridges)
library(ggrepel)
library(RColorBrewer)
library(jcolors)
library(scales)
library(igraph)
library(patchwork)
library(readxl)
library(unikn)     
library(gridExtra)
library(grid)
library(data.table)
library(plyr)
library(inline) 
library(ctc)
library(gplots)  # Visual plotting of tables
library(ggplot2)
library(nortest)
library(dplyr)
library(devtools)
library(bedr)
library(rtracklayer)
library(broom)
library(reshape2)
library(gridBase)
library(formattable)
library(tidyr)
library(glmmADMB)
library(openxlsx)
library(MASS)
library(ggfortify)
library(pals)
library(pheatmap)
library(viridis) 
library(scales)
library(tilingArray)
library(R.filesets)
library(effsize)
library(nlme)
library(broom.mixed)
brainregions <- setNames(c("#1B9E77" ,"#D95F02" ,"#7570B3", "#E6AB02" ,"#666666" ), c('Cerebellum','Entorhinal cortex', 'Hippocampal CA1','Hippocampal CA3', 'Temporal cortex'))
